Client: fix broken sorting in ins_file_meta()
Problem which the fix solves:
When the second file is being written on the client,
it crashes with segfault in _chunk_write() because bsearch()
returns NULL.
The issue is with ins_file_meta() that doesn't add a new fattr entry
in acsending order which is required by bsearch().
This fix patches ins_file_meta() so it always inserts elements properly.